Add: an acceptance (or integration) test suite for scap
Also, add scripts/check to run all the tests.

Also, run the tests during a build of the Debian package, if Subplot
is installed. Subplot is not yet in Debian itself, so there is no hard
build-dependency on it.

All the runtime dependencies are added as build dependencies, so scap
can run during Debian package build.

Change-Id: I523e72ba1baae19fca3b84d312c400d94c3c0d06